Adds a new dependency to test fips specific revision importing feature.
The new dependency emulates an API change in a project in progress that may
break projects that depends on it. This project depends on a specific
version of the dep3 project, the project in question does not uses tagging,
so the only way to reference the good version is by using a commit SHA1-ref.